pyproject.toml
[tool.poetry]
name = "modernexcelreports"
version = "0.1.0"
description = "Build Visual Excel(s) Tab(s) from Python"
authors = ["Shivam Anand <shivam01anand@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
pandas = "^1.4.1"
Pillow = "^9.0.1"
XlsxWriter = "^3.0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
